// src/interfaces/DetailsListBasicExampleItem.type.ts
export interface IDetailsListBasicExampleItem {
  ID: number;
  Title: string;
  Entity: [];
  VendorStatus: string;
  GxP_x0020_Suppliers: string;
  Nature_x0020_of_x0020_Product: string;
  HCP_x002f_PAG_x002f_PO_x0020_Rep: string;
  Healthcare_x0020_Organization: string;
  Candidate_x0020_reimbursement: string;
  Payment_x0020_Method: string;
  Terms_x0020_of_x0020_Payment: string;
  Billing_x0020_Currency: string;
  Tax_x0020_Status: string;
  Tax_x0020_ID: string;
  Contact_x0020_Name: string;
  Contact_x0020_Email: string;
  Phone_x0020_Number: any;
  Fax_x0020_Number: any;
  Vendor_x0020_website: string;
  Address_x0020_Line_x0020_1: string;
  Address_x0020_Line_x0020_2: string;
  Address_x0020_Line_x0020_3: string;
  Country: string;
  State: string;
  City: string;
  Zip: string;
  HCP_x002f_PAG_x002f_PO_x0020_Nam: string;
  HCP_x002f_PAG_x002f_PO_x0020_Cou: [];
  Created: any;
  RequestorsName: any;
  IsCheck: any;
}